pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "simpletrace"
version = "0.1.0"
description = "A small torch training trace tool"
readme = "README.md"
requires-python = ">=3.7"
license = { text = "MIT" }
authors = [
  { name = "Gaosiyuan", email = "siyuan-gao@outlook.com" },
]

classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "torch>=2.0.0",
  "pyyaml",
]

[project.urls]
Homepage = "https://github.com/birdswimming/SimpleTrace"

[project.scripts]
simpletrace-profile = "simpletrace.profile_enable:profile_enable"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["simpletrace*"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.yaml", "*.sh"]